Moved definition of compiler flags so they become active. This fixes the lack of debug info in SLC6. Fixed superflous semicolon detected by -pedantic Fixed sign correctness in various places. Removed the use of LL suffix in 64 bits contants (they are not required). Some documentation clarifications.

+++ Exception/Exception.cc
@@ -80,7 +80,7 @@ Tape::Exceptions::Backtrace::Backtrace() {
     m_trace = "";
   else {
     std::stringstream trc;
-    for (int i=0; i<depth; i++) {
+    for (size_t i=0; i<depth; i++) {
       std::string line(strings[i]);
       /* Demangle the c++, if possible. We expect the c++ function name's to live
        * between a '(' and a +
